Health chatbot/requirements.txt
streamlit
google-generativeai
SpeechRecognition
python-dotenv
transformers
pyttsx3
torch
